+++ src/features/devtools/data/devToolsSagas.ts
@@ -2,7 +2,6 @@ import { call, put } from 'redux-saga/effects';
 import type { SagaIterator } from 'redux-saga';
 
 import { DevToolsActions } from 'src/features/devtools/data/devToolsSlice';
-import { waitForSelector } from 'src/utils/pdf';
 
 export function* previewPdfSaga(): SagaIterator {
   yield put(DevToolsActions.setPdfPreview({ preview: true }));
@@ -11,4 +10,15 @@ export function* previewPdfSaga(): SagaIterator {
     window.print();
   }
   yield put(DevToolsActions.setPdfPreview({ preview: false }));
+}
+
+async function waitForSelector(selector: string, timeOut = 5000) {
+  const start = performance.now();
+  while (document.querySelector(selector) === null) {
+    if (performance.now() - start > timeOut) {
+      return null;
+    }
+    await new Promise((resolve) => requestAnimationFrame(resolve));
+  }
+  return document.querySelector(selector);
 }
